+++ insider/src/main/java/com/kh/insider/configuration/InterceptorConfiguration.java
@@ -32,6 +32,7 @@ public class InterceptorConfiguration implements WebMvcConfigurer {
 //						"/member/nickCheck",
 //						"/member/sendMail",
 //						"/member/passwordChange",
+//						"/member/suspension",
 //						"/static/**"
 //						);
 		//Admin 인터셉터
