+++ insider/src/main/resources/mybatis/mapper/member-suspension-mapper.xml
@@ -15,7 +15,7 @@
 		update member_suspension set member_suspension_status=1, member_suspension_lift_date=sysdate where member_no=#{memberNo} and member_suspension_status=0 
 	</update>
 	<update id="addSuspension">
-		update member_suspension set member_suspension_status=0, member_suspension_times=member_suspension_times+1, member_suspension_days=#{memberSuspensionDays}, member_suspension_lift_date=sysdate+#{memberSuspensionDays} where member_no=#{memberNo}
+		update member_suspension set member_suspension_status=0, member_suspension_times=member_suspension_times+1, member_suspension_days=#{memberSuspensionDays}, member_suspension_lift_date=sysdate+#{memberSuspensionDays}, member_suspension_content=#{memberSuspensionContent} where member_no=#{memberNo}
 	</update>
 	<update id="updateReportContent">
 		UPDATE MEMBER_SUSPENSION SET member_suspension_content=#{reportListContentAfter} WHERE member_suspension_content=#{reportListContentBefore}
