+++ insider/src/main/webapp/WEB-INF/views/member/mypage.jsp
@@ -871,7 +871,7 @@
           <div class="modal" tabindex="-1" role="dialog" id="followerModal"
                             data-bs-backdrop="static"
                             ref="followerModal" @click.self="followerModalHide">	
-            <div class="modal-dialog" role="document">
+            <div class="modal-dialog d-flex justify-content-center align-items-center" role="document" style="height:80%">
                    <div class="modal-content" style="max-width:400px; min-height:400px; max-height:400px;">
                        <div class="modal-header text-center" style="display:flex; justify-content: center; height:50px;">
 							<h5 class="modal-title">팔로워</h5>
@@ -971,7 +971,7 @@
           <div class="modal" tabindex="-1" role="dialog" id="followModal"
                             data-bs-backdrop="static"
                             ref="followModal" @click.self="followModalHide">
-             <div class="modal-dialog" role="document">
+             <div class="modal-dialog d-flex justify-content-center align-items-center" role="document" style="height:80%">
     <div class="modal-content" style="max-width:400px; min-height:200px; max-height:400px;">
       <div class="modal-header text-center" style="display:flex; justify-content: center; height:50px;">
         <h5 class="modal-title" >팔로잉</h5>
